ICT/requirements.txt
streamlit==1.38.0
pyshorteners
qrcode[pil]
pillow
pandas
numpy